first commit
first commit

+++ sources/OAL_Source.cpp
@@ -673,7 +673,7 @@ void cOAL_Source::SetElapsedTime(double afTime)
 	case eOAL_AudioDataType_Stream:
 		{
 			cOAL_Stream* pStream = (cOAL_Stream*)mpAudioData;
-			pStream->Seek((float)afTime/mpAudioData->GetTotalTime());
+			pStream->Seek((float)afTime/(float)mpAudioData->GetTotalTime());
 		}
 		break;
 	}
